MXS-3084: Fix forced service destruction
The iteration over the targets used a reference to the container that was
being modified. Removal of items from a std::vector invalidates iterators
past the item being removed which leads to problems after the first target
is erased.

The removal of targets from services isn't necessary and the REST API code
can just bypass these checks.